+++ src/app/globals.css
@@ -31,7 +31,6 @@ main {
   min-height: 100vh;
   flex-direction: column;
   align-items: center;
-  justify-content: space-between;
 }
 
 @layer utilities {
