+++ src/app/page.tsx
@@ -3,7 +3,7 @@ import TrendingBanner from "@/components/TrendingBanner/TrendingBanner"
 export default async function Home() {
   return (
     <main>
-      <TrendingBanner></TrendingBanner>
+      <TrendingBanner />
     </main>
   )
 }
